feat: rename new hook back to original name
