+++ apps/web/partials/editor/table-node.tsx
@@ -4,8 +4,7 @@ import { ErrorBoundary } from 'react-error-boundary';
 import * as React from 'react';
 import { useMemo } from 'react';
 
-import { TableBlockStoreProvider } from '~/core/state/table-block-store';
-import { TableBlockProvider } from '~/core/state/table-block-store-v2';
+import { TableBlockProvider } from '~/core/state/table-block-store';
 import { useTypesStore } from '~/core/state/types-store/types-store';
 import { GeoType } from '~/core/types';
 
@@ -85,11 +84,9 @@ function TableNodeChildren({
         </>
       }
     >
-      <TableBlockStoreProvider spaceId={spaceId} entityId={entityId} selectedType={selectedType}>
-        <TableBlockProvider spaceId={spaceId} entityId={entityId} selectedType={selectedType}>
-          <TableBlock spaceId={spaceId} />
-        </TableBlockProvider>
-      </TableBlockStoreProvider>
+      <TableBlockProvider spaceId={spaceId} entityId={entityId} selectedType={selectedType}>
+        <TableBlock spaceId={spaceId} />
+      </TableBlockProvider>
     </ErrorBoundary>
   );
 }
